+++ packages/profilecard/stories/examples/ProfilecardTrigger.jsx
@@ -1,11 +1,8 @@
 import React from 'react';
-import {
-  AkProfilecardTrigger,
-} from '@atlaskit/profilecard';
+import { AkProfilecardTrigger } from '@atlaskit/profilecard';
+import { getMockProfileClient } from '../util';
 
-import MockProfileClient from '../story-data';
-
-const mockClient = new MockProfileClient();
+const mockClient = getMockProfileClient();
 
 const canvasStyles = {
   paddingBottom: '360px',
@@ -26,6 +23,24 @@ export default (
       ]}
     >
       <strong>Hover Over Me</strong>
+    </AkProfilecardTrigger>
+
+    &hellip;
+
+    <AkProfilecardTrigger
+      position="bottom left"
+      cloudId="DUMMY-CLOUDID"
+      userId="2"
+      resourceClient={mockClient}
+      trigger="click"
+      actions={[
+        {
+          label: 'View profile',
+          callback: () => {},
+        },
+      ]}
+    >
+      <strong>Click me</strong>
     </AkProfilecardTrigger>
   </div>
 );